rename files ess-upload-results to ess-list-encs component
