+++ src/app/features/exchange-set/exchange-set.module.ts
@@ -6,7 +6,7 @@ import { ExchangeSetRoutingModule } from './exchange-set-routing.module';
 import { ExchangeSetComponent } from './exchange-set.component';
 import { EssUploadFileComponent } from './ess-upload-file/ess-upload-file/ess-upload-file.component';
 import { EssAddSingleEncsComponent } from './ess-add-single-encs/ess-add-single-encs.component';
-import { EssListEncsComponent } from './ess-upload-results/ess-list-encs.component';
+import { EssListEncsComponent } from './ess-list-encs/ess-list-encs.component';
 
 @NgModule({
   declarations: [
